Expose getOperationType(operation) on GraphQLSchema
We're finding we frequently do switch statements on an OperationDefinition's operation to pull the right type out of the schema. This hardens that approach: if for some reason we added a new operation kind, you could just update this function, rather than all the callsites that match on the operation type.