Pretty format variable definitions for both operations and fragments.
When there is more than a couple argument values, pretty formatting
isn't doing a good job.

Instead I would like to take the following approach:

When there is one argument, print as is, meaning:

  query Q ($arg: Type = Default) {

When there is more than one do the following:

  query Q (
    $arg1: Type = Default,
    $arg2
  ) {

An alternative would be to have closing bracket be in the same line
as final argument. I am fine with either, as changes which introduce
another argument on last position always touch two lines - traiiling
comma will have to be added.