Change printError() to mark location before using ↓ instead of ^ after.
Queries, particularly when compressed, can have a long line length. In narrow viewports (small Terminal windows, etc.) this results in the query wrapping over multiple lines.

Previously the error line location was marked on a line after with sufficient padding and a ^ character. The problem is, when the line above wraps the marker points up at an incorrect location.

By putting the marker on a line before, not after, there is less chance it will point at the wrong location on the following line due to wrapping. At least if the error occurred anywhere in the first line wrap (fairly likely) marker will be correct.

Also, the ↓ symbol on the line before is more intuitive because people read from top to bottom and the symbol itself is more semantic.