fix: change OperationTypeNode to enum
Changes OperationTypeNode to enum instead of a
union type. This allows developers to use the type
as `OperationTypeNode.QUERY` instead of risking to
msispell the string 'query'.

Example:

Instead of having to write

`operation === 'query'`

you can instead write

`operation === OperationTypeNode.QUERY`

This should not be a breaking change, since the string
version should still work.