introspection: Allow interfaces to be undefined for interfaces
The old code was using strict equally check for null to check if the introspected field `interfaces` was not defined, which means that it would fail in case `interfaces` was set to undefined. Older GraphQL implementations will not send the `interfaces` field when introspecting an interface, and thus would fail with `Introspection result missing interfaces`.