feat: add CustomExecutor option to graphql, execute, and subscribe
allows customization of the execution algorithm by overriding any of the
protected members of the now exported internal Executor class.

Reference:

https://github.com/graphql/graphql-js/pull/3163#issuecomment-859546546

Note:

This class is exported only to assist people in implementing their own
executors without duplicating too much code and should be used only as
last resort for cases requiring custom execution or if certain features
could not be contributed upstream.

It is still part of the internal API and is versioned, so any changes
to it are never considered breaking changes. If you still need to
support multiple versions of the library, please use the `versionInfo`
variable for version detection.