Align calls of buildExecutionContext
`buildExecutionContext` used to throw an error on failure, but was changed
to return an array of errors (enabling the function to return multiple
errors encountered during variable coercion).

As the function is no longer designed to throw, it no longer needs to be
wrapped in a try block when used within the `subscribe` function. This
change aligns the use of `buildExecutionContext between the `subscribe`
and `execute` functions.