Validate variables are known types
Adds a check to the validation rule VariablesAreInputTypes to ensure
that referenced types exist at all before checking they are input types.

This check is not explicitly described in the specification for validating
variables: http://spec.graphql.org/draft/#sec-Validation.Variables
I am assuming this is simply an oversight and the rule is implicit.
Still, it is valuable to check for it in order to provide a proper error to clients.