Add option printDirective to printSchema
Resolves https://github.com/graphql/graphql-js/issues/2020

I initially thought about implementing this as https://github.com/graphql/graphql-js/issues/2020#issuecomment-866798901,
but really could not think of a good use case where
a directive should be printed in a modified way.

Thus, I went for a simpler signature:

```ts
shouldPrintDirective?: (directiveNode: DirectiveNode) => boolean
```

This is only a partial implementation for now, there are more possible
locations in a schema that can have a directive. Before I finish the
implementation, I would like to validate the direction is right.