GraphQLError: switch constructor overload order
The GraphQLError constructor has two overloads. The GraphQLErrorArgs one
is recommended and the one with a bunch of position arguments is
deprecated.

When typechecking, TypeScript chooses the *first* matching overload (see
eg
https://github.com/microsoft/TypeScript/issues/1860#issuecomment-72154737).
Passing just a message argument matches both overloads, so this
currently displays as deprecated.

This PR reverses the order so that passing just a message does not
display as deprecated.