github-action-bot: use URL for benchmark report instead of embedding it into comment
Currently benchmark report is posted directly to comment, but because of
terminal color codes it look unreadable. Sadly, github doesn't provide
any mechanism to add colors to the comment text (even in code blocks).
The best solution I can think of is to link to GH own logs to show
benchmark report