refactor: DRY with context building and *Impl functions
The `execute`/`executeImpl` and `createSourceEventStream`/`createSourceEventStreamImpl` functions follow the same basic pattern of building the contet and using it to run a function. This PR extracts that pattern into a separate function.

For good measure, the same pattern in applied to the soon-to-be-deprecated `subscribe` function.

Hheavier refactoring is on the way from @IvanGoncharov (see https://github.com/graphql/graphql-js/pull/3639#pullrequestreview-1013806199), but in the meantime, this consolidates the common pattern without any breaking changes.