getSubscriber does not return an AsyncGenerator
Returns an object that conforms to the AsyncGenerator type signature, but the behavior of the properties do not conform to the AsyncGenerator specification(https://tc39.es/ecma262/#sec-properties-of-asyncgenerator-prototype).

In particular, with async generators, all promises returned by next, return, and throw resolve in call order.

TS does not include a type that is AsyncGeneratorLike, and AsyncIterableIterator has an optional return, while here return is present, so the best things seems to be to just let TS infer the type.