defer/stream: fix flattenAsyncIterable concurrency
Fixes the bug demonstrated in #3709 (which has already been incorporated
into the defer-stream branch).

This fix is extracted from #3703, which also updates the typing and API
around execute.

This particular change doesn't affect the API, other than making the
`subscribe` return type more honest, as its returned generator could
yield AsyncExecutionResult before this change as well. (The reason the
previous version built is because every ExecutionResult is actually an
AsyncExecutionResult; fixing that fact is part of what #3703 does.)