incrementalDelivery: fix iterable streaming with errors
When iterables (not asyncIterables) are streamed, and completeValue throws an error, we currently send undefined rather than null, due to a typo.

This change corrects the typo, and also modifies the test intended to check this case to correctly check a streamed iterable, rather than an asyncIterable.

Code coverage was complete prior to this change, because the line involving this change was reached in the case where the error was non-recoverable, albeit with the line throwing.