fix(incrementalDelivery): add additional tests
When a non-null list is streamed and completeValue for a list item
fails, the error should bubble up to the list itself and no further
payloads should be sent.

Addition of tests within this change demonstrates that this is currently
the case only when the field resolver for the list returns an iterable,
but not when it returns an async iterable. In the latter case, the null
correctly bubbles to the list itself for the given payload (`items:
null` rather than `items: [null`), but further payloads are sent.