workflows: fix and cleanup permissions
Context: #3751 added permissions to workflows.
It broke some of the workflows and was partially fixed in 3759
This PR attempt to fix rest of workflows and also cleanup permissions in general.
Assignining granular permissions is error-prone especially if we need to account of top-level workflow's permissions.
This PR attempt to solve this by explicitly listing permissions on each individual job.
It slightly increase verbosity for some workflows but is more maintanable by keeping permissions closer to steps that require them.