polish: fix test
I introduced use of expect.fail() in #3760 mistakenly believing that expect.fail() will always cause the test to fail.

Actually, expect.fail() just throws an error, and that error is wrapped is wrapped by GraphQL and then filtered, so the lines are reached.

Fortunately, it's not a problem that the line is reached, as long as it is filtered correctly. The number of calls to next() may be more than 1 (in our case it is 2) depending on the # of ticks that it takes for the deferred fragment to resolve.

So the test as written is intending to be overly strict, but is broken and so functions as desires. This commit makes no change to the test semantics, but changes the comment, the coverage ignore statement, and removes the confusing use of expect.fail, so that the test semantics are more clearly apparent.