ESLint: switch to '.eslintrc.cjs' as config file
Context: As disscussed in #3849 ESLint team is planning to deprecate the current config format and switch to:
https://eslint.org/docs/latest/use/configure/configuration-files-new
But it requires that all plugins that we use to support this format.

In my PoC attempt to use this new format biggest change is comming from converting YAML file into JS, so I decided to extract it into separate commit.
Note: beyound JS convertion I did a purely stylistic cleanup of `TODO` comments, beyound that this commit is pure convertion of YAML into JS.