Drop support for node14
Motivation: technically node14 end of life will happen in two months so we can still support it but it creates problems with #3791
Node14 doesn't support `AbortSignal` and since we want to allow resolvers to be canceled we need to expose this object through our public API (inside `info` argument).
Potentialy this and all the other issues (e.g. how to test this feature on node14) can be worked around but complicates solution.
So I think, it worth to deprecate node14 two months earlier in order to unblock #3791 especially since we still in alpha stage.

Less critical but still significant problem with node14 is that it ships with npm6 which doesn't support default format of the lockfile
used by npm9 and this make updating dependencies trickier.