refactor(incremental): set pending result id during execution
This makes id an immutable property of each pending result, reducing the need to track these ids / lazily create them within the publisher.

Empty and filtered deferred fragments will lead to "missing" ids and id ordering will now depend on timing/early execution. The id by the specification are opaque strings so that this is not a concern.