standardize error messages prior to introducing schema coordinates
extracted from #3808

PR #3808 uses schema coordinates to improve GraphQL-JS error messages. To reduce the size of the PR, this commit standardizes error messages according to the general pattern that will be introduced with schema coordinates without introducing the coordinates themselves, in the hopes of aiding review of the later PR.

Parenthetically, extracting these changes out of #3808 demonstrates how the use of schema coordinates improves our error messages. For example, it is difficult when printing an error message regarding an argument to easily look up the field of that argument, and, if it a nested input field, the full field path. It is similarly sometimes difficult to know the type of the field or argument triggering the error. Embedding schema coordinates within GraphQL-JS makes this trivial. This PR simply omits the changes from #3808 that are directly enabled by the use of schema coordinates. #3808 can then be rebased on top of this PR to demonstrate the code changes and the direct improvement in the various error messages.