Fix/optimize value coercion check for OneOf type (#4181)
Where a JavaScript value is coerced to a GraphQL OneOf Input Type, a
null value should only be reported when the "one of" condition is
satisfied. The code block starting at line 158
[here](https://github.com/graphql/graphql-js/blob/1dbdadc6f46d2c97c71c7a4f5c61a2c75d1536ae/src/utilities/coerceInputValue.ts#L158)
that accesses `keys[0]` should not be executed if `keys` is empty or
contains more than one item.

The PR fixes this by adding an else statement. Alternatively, the
"if/else" branches could be reversed, and the `keys.length !== 1` check
should become a `keys.length === 1` check.