Convert website to nextra (#4240)
This converts the existing website to nextra just like
https://github.com/graphql/graphql.github.io. This is a first step in
moving the documentation here and having a redirect from graphql.org to
graphql-js.org.

Not sure yet why codecov started failing 😅 when I run `testonly:cover`
locally it tells me we are 100% covered.

WDYT about isolating the dependencies for the website in that folder? As
seen in
https://github.com/graphql/graphql-js/pull/4240/commits/9c7d615bfe9669c01081156e32bc25c7b24f795c
this prevents the weird CI leaks that we're seeing

Resolves https://github.com/graphql/graphql-js/issues/4200

---------

Co-authored-by: Yaacov Rydzinski <yaacovCR@gmail.com>