fix sidebar instability
when navigating to https://www.graphql-js.org/api-v16/graphql/ using the top menu bar, the sidebar content expands with the individual mdx files from the api-v16 folder, as if the api/v16 option had been in the sidebar in the first place.

This PR adds the API/v16 to the sidebar (in addition to the placement within the menu bar) to avoid this instability.