Add support for @experimental_disableErrorPropagation (#4348)
~This pull request adds support for `@onError(action: NULL)` to disable
error propagation for error aware clients:~

This pull request adds support for
`@experimental_disableErrorPropagation` to disable error propagation for
error aware clients:

```graphql
"""
Disables error propagation.
"""
directive @experimental_disableErrorPropagation on QUERY | MUTATION | SUBSCRIPTION
```

I'm not super used to write TypeScript, feel free to amend the PR as
needed but I figured it'd be good to have.

The logic is unconditional. The matching
[graphql-java](https://github.com/graphql-java/graphql-java/pull/3772)
PR has a specific opt-in flag so that it's not enabled by accident in
the very unlikely event that a schema already contains a matching
directive. Let me know if this is an issue.

Many thanks @JoviDeCroock for pointing me in the right direction 🙏

See https://github.com/graphql/nullability-wg/discussions/85
See https://github.com/graphql/graphql-spec/pull/1050

---------

Co-authored-by: Jovi De Croock <decroockjovi@gmail.com>