fixed wrong variable name running-an-express-graphql-server.mdx
Issue
The server throws an error because the root variable is not defined. The correct variable name should be rootValue, which was previously declared in the code.

Fix
Updated rootValue in the createHandler function to use the correct variable name.

Changes
Replaced root with rootValue in the GraphQL handler configuration.