test: actually test use of AbortSignal in resolver
The previous test didn't really make use of the fact that the
AbortSignal could be used in the resolver: the only use it made of it
was to call `signal.throwIfAborted` *after* the cancellable promise was
already cancelled. The "This operation was aborted" message that shows
up in the GraphQL response actually came from the cancellable promise,
not the throwIfAborted call. You can see that because if you just
replace `throwIfAborted` with throwing another error (as this commit
does), the test still passed.

Instead, actually make use of the AbortSignal API to observe the abort
explicitly.