avoid re-visiting deferred fragment spreads
This PR changes the fragment spread visiting logic to avoid re-visiting fragment spreads whenever possible.

Previously, fragment spreads that have been visited in a non-deferred capacity were not revisited, but fragment spreads that had been visited in a deferred capacity could be revisited.

This PR updates the logic so that fragment spreads that had been visited in a deferred capacity are not revisited in a deferred capacity.

See https://github.com/graphql/graphql-spec/pull/1045#issuecomment-3085263996