refactor: extract Queue from IncrementalPublisher and IncrementalGraph
Queue is a batching async-aware iterator-like protocol meant to be consumed in the following (and only the following) pattern:

>  let batch;
>  if ((batch = queue.currentBatch()) !== undefined) {
>    doSomethingWithBatch(batch);
>  }
>  while ((batch = await queue.nextBatch) !== undefined) {
>    doSomethingWithBatch(batch);
>  }

A `push()` methods is provided in the Queue constructor (a la repeaters, see https://repeater.js.org/) so that `push()` can be private to the code that constructs the Queue (although it can be saved to be passed to other code).

A `stop()` method is also provided as a the second argument to the constructor for convienence, but it is also available on the queue object itself so that it can be called by the executor or by the consumer. So this works:

>  const queue = new Queue(
>    async (push, stop) => {
>      push(1);
>      push(2)
>      await Promise.resolve();
>      push(3);
>      stop();
>    },
>  );
>
>  const batch1 = Array.from(queue.nextBatch()); // batch1 = [1, 2]
>  const batch2 = Array.from(await queue.nextBatchAsync()); // batch2 = [3]

as does this:

>  let push;
>  const queue = new Queue(
>    (_push) => {
>      push = _push;
>    },
>  );
>
>  push(1);
>  push(2);
>
>  const batch1 = Array.from(queue.nextBatch()); // batch1 = [1, 2]
>
>  const batch2Promise = queue.nextBatchAsync();
>
>  await Promise.resolve();
>  push(3);
>  queue.stop();
>
>  const batch2 = await batch2Promise; // batch2 = [3]

Note: concurrent calls to `currentBatch()` and `nextBatch` will return the same batch and are not encouraged.

A `toAsyncIterable(mapFn)` method transforms coalesces each batch of items into a single value (or undefined if not value is to be emitted for the batch).

Using queues, we are able to remove all logic for handling the implicit queue from IncrementalPublisher, retaining only the `_handleCompletedBatch()`, while adding only the required `push()` and `stop()` calls within the IncrementalGraph.

Tests do not change, except that `.return()` and `.throw()` (but not `next()` have an extra tick secondary to the additional layers of `withCleanup()`, so that the tests required slight adjustment.